requirements.txt
Django==4.2.11
django-filter==23.5
django-cors-headers==4.3.1
django-ckeditor==6.7.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
Pillow==10.2.0
python-dotenv==1.0.1
stripe==8.4.0
razorpay==1.4.1
paypalrestsdk==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.6
gunicorn==21.2.0
whitenoise==6.6.0
django-storages==1.14.2
boto3==1.34.51 